sistema_de_informes/services/api-gateway/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
websockets==12.0
